+++ src/ingest/index.ts
@@ -1,4 +1,3 @@
 export { Collection } from "./collection";
 export { Cursor } from "./cursor";
 export { Fetcher } from "./fetcher";
-export { Publisher } from "./publisher";
